Add JSON dump command
