+++ modules/codegen-cli/src/smithy4s/codegen/cli/CodegenCommand.scala
@@ -17,32 +17,14 @@
 package smithy4s.codegen.cli
 
 import cats.data.Validated
-import cats.data.ValidatedNel
 import cats.syntax.all._
-import com.monovore.decline.Argument
 import com.monovore.decline.Command
 import com.monovore.decline.Opts
-import os.Path
 import smithy4s.codegen.CodegenArgs
 
-import java.nio.file
+import Options._
 
 object CodegenCommand {
-
-  implicit val osPathArg: Argument[os.Path] = new Argument[os.Path] {
-    def defaultMetavar: String = "path"
-    def read(string: String): ValidatedNel[String, Path] =
-      implicitly[Argument[file.Path]].read(string).andThen { path =>
-        try {
-          if (path.isAbsolute()) Validated.validNel(os.Path(path))
-          else Validated.validNel(os.pwd / os.RelPath(path))
-        } catch {
-          case e: Throwable =>
-            Validated.invalidNel(e.getMessage() + ":" + string)
-        }
-      }
-
-  }
 
   val outputOpt =
     Opts
@@ -96,38 +78,6 @@ object CodegenCommand {
       .map(_.split(',').toSet)
       .orNone
 
-  val specsArgs = Opts
-    .arguments[os.Path]()
-    .mapValidated(
-      _.traverse(path =>
-        if (os.exists(path)) Validated.valid(path)
-        else Validated.invalidNel(s"$path does not exist")
-      )
-    )
-    .orNone
-    .map {
-      case Some(value) => value.toList
-      case None        => List.empty
-    }
-
-  val repositoriesOpt: Opts[Option[List[String]]] =
-    Opts
-      .option[String](
-        "repositories",
-        "Comma-delimited list of repositories to look in for resolving any provided dependencies"
-      )
-      .map(_.split(',').toList)
-      .orNone
-
-  val dependenciesOpt: Opts[Option[List[String]]] =
-    Opts
-      .option[String](
-        "dependencies",
-        "Comma-delimited list of dependencies containing smithy files"
-      )
-      .map(_.split(',').toList)
-      .orNone
-
   val options =
     (
       outputOpt,
@@ -156,8 +106,8 @@ object CodegenCommand {
       }
 
   val command = Command(
-    "smithy4s",
+    "generate",
     "Generates scala code and openapi-specs from smithy specs"
-  )(options)
+  )(options.map(Smithy4sCommand.Generate))
 
 }
